moved Home and HomeViewModel to home package
