+++ app/src/main/java/com/chydee/notekeeper/home/HomeViewModel.kt
@@ -1,4 +1,4 @@
-package com.chydee.notekeeper
+package com.chydee.notekeeper.home
 
 import androidx.lifecycle.ViewModel
